Public tests for HubLifetimeManager
