+++ test/Microsoft.AspNetCore.SignalR.Tests.Utils/TestClient.cs
@@ -14,7 +14,12 @@ using Microsoft.AspNetCore.SignalR.Protocol;
 
 namespace Microsoft.AspNetCore.SignalR.Tests
 {
-    public class TestClient : ITransferFormatFeature, IConnectionHeartbeatFeature, IDisposable
+#if TESTUTILS
+    public
+#else
+    internal
+#endif
+    class TestClient : ITransferFormatFeature, IConnectionHeartbeatFeature, IDisposable
     {
         private readonly object _heartbeatLock = new object();
         private List<(Action<object> handler, object state)> _heartbeatHandlers;
